% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_module.R
\name{assign_gene}
\alias{assign_gene}
\title{Assign genes to groups.}
\usage{
assign_gene(gene_data, reg_group_table)
}
\arguments{
\item{gene_data}{A n by p matrix of expression data with rows for genes and columns for samples.}

\item{reg_group_table}{A g by p matrix with i-th row for the grouping of samples of i-th group.}
}
\value{
A list of length n, which are the indices of groups for the genes.
}
\description{
Assign genes to groups that has highest log likelihood based on the grouping of samples.
}
\examples{
reg_group_table=matrix(c(rep(0,3),rep(1,9),rep(0,9),rep(1,3)),nrow=2,byrow=TRUE)
assign_gene(gene_data = matrix(rnorm(100*12),100,12),reg_group_table=reg_group_table)
}
