% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_module.R
\name{build_split_table}
\alias{build_split_table}
\title{Build split table by K-means heuristicly.}
\usage{
build_split_table(X)
}
\arguments{
\item{X}{A n by p matrix as input.}
}
\value{
A n by p matrix with each column consists of 3 clusters: -1 for low, 0 for mid and 1 for high
}
\description{
Build split table by K-means with 3 cluster centers for each column of X
}
\examples{
split_table <- build_split_table(matrix(rnorm(5*10),5,10))
}
