% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_module.R
\name{calc_likelihood_score}
\alias{calc_likelihood_score}
\title{Calculate Gaussian Likelihood score.}
\usage{
calc_likelihood_score(x, labels)
}
\arguments{
\item{x}{A n by p matrix.}

\item{labels}{A vector of length n, indicating the group of rows.}
}
\value{
The sum of log likelihood score of each group on each column.
}
\description{
Calculate Gaussian Likelihood score.
}
\examples{
calc_likelihood_score(x = matrix(rnorm(50*100),50,100), labels = c(rep(1,20),rep(2,30)))
}
