% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnet_r.R
\name{get_correlation_list}
\alias{get_correlation_list}
\title{Calculate correlation within each group.}
\usage{
get_correlation_list(x, labels)
}
\arguments{
\item{x}{A n by p matrix.}

\item{labels}{A vector of length n, indicating the group of rows.}
}
\value{
An array of Pearson correlation coefficient for each row, rows belong to the same group have same values.
}
\description{
Calculate Pearson correlation coefficient within each group.
}
\examples{
get_correlation_list(x = matrix(rnorm(50*100),50,100), labels = c(rep(1,20),rep(2,30)))
}
