% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot the regression tree.}
\usage{
plot_tree(gnet_result, group_idx)
}
\arguments{
\item{gnet_result}{Results returned by gnet().}

\item{group_idx}{Index of the module..}
}
\value{
None
}
\description{
Plot the regression tree given the index of a module.
}
\examples{
set.seed(1)
exp_data <- matrix(rnorm(100*12),100,12)
tf_list <- paste0('TF',1:10)
rownames(exp_data) <- c(tf_list,paste0('gene',1:(nrow(exp_data)-length(tf_list))))
colnames(exp_data) <- paste0('condition_',1:ncol(exp_data))
se <- SummarizedExperiment::SummarizedExperiment(assays=list(counts=exp_data))
gnet_result <- gnet(se,tf_list)
plot_tree(gnet_result,group_idx=0)
}
