% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_module.R
\name{gnet}
\alias{gnet}
\title{Run GNET2}
\usage{
gnet(input, reg_names, init_group_num = 20, max_partition_level = 4,
  cor_cutoff = 0.9, min_divide_size = 3, min_group_size = 2,
  max_iter = 5)
}
\arguments{
\item{input}{A SummarizedExperiment object, or a p by n matrix of expression data of p genes and n samples, for example log2 RPKM from RNA-Seq.}

\item{reg_names}{A list of potential upstream regulators names, for example a list of known transcription factors.}

\item{init_group_num}{Initial number of function clusters used by the algorithm.}

\item{max_partition_level}{max_partition_level Maximum partition level in the tree.}

\item{cor_cutoff}{Cutoff for within group Pearson correlation coefficient, if all data belong to a node have average correlation greater or equal to this, the node would not split anymore.}

\item{min_divide_size}{Minimum number of data belong to a node allowed for further split of the node.}

\item{max_iter}{Maxumum number of iterations allowed if not converged.}

\item{min_divide_size}{Minimum number of genes allowed in a group.}
}
\value{
A list of expression data of genes, expression data of regulators, within group score, table of tree structure and final assigned group of each gene.
}
\description{
Build regulation modules by iteratively perform TF assigning and Gene assigning, until the assignment of genes did not change, or max number of iterations reached.
}
\examples{
set.seed(1)
exp_data <- matrix(rnorm(1000*12),1000,12)
tf_list <- paste0('TF',1:100)
rownames(exp_data) <- c(tf_list,paste0('gene',1:(nrow(exp_data)-length(tf_list))))
colnames(exp_data) <- paste0('condition_',1:ncol(exp_data))
se <- SummarizedExperiment(assays=list(counts=exp_data))
gnet_result <- gnet(se,tf_list,init_group_num,max_partition_level,cor_cutoff,min_divide_size,min_group_size,max_iter)
}
