% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnet_r.R
\name{kneepointDetection}
\alias{kneepointDetection}
\title{Knee point detection.}
\usage{
kneepointDetection(vect)
}
\arguments{
\item{vect}{A list of sorted numbers.}
}
\value{
The index of the data point which is the knee.
}
\description{
Detect the knee point of the array.
}
\examples{
kneepointDetection(sort(rnorm(100),T))
}
