% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_gene_group}
\alias{plot_gene_group}
\title{Plot a module}
\usage{
plot_gene_group(gnet_result, group_idx)
}
\arguments{
\item{gnet_result}{Results returned by gnet().}

\item{group_idx}{Index of the module.}
}
\value{
None
}
\description{
Plot the regulators module and heatmap of the expression inferred downstream genes for each sample. 
It can be interpreted as two parts: the bars at the top shows how samples are splited by the 
regression tree and the heatmap at the bottom shows how downstream genes are regulated by each 
subgroup determined by the regulators.
}
\examples{
set.seed(1)
init_group_num = 10
init_method = 'boosting'
exp_data <- matrix(rnorm(100*12),100,12)
reg_names <- paste0('TF',1:10)
rownames(exp_data) <- c(reg_names,paste0('gene',1:(nrow(exp_data)-length(reg_names))))
colnames(exp_data) <- paste0('condition_',1:ncol(exp_data))
se <- SummarizedExperiment::SummarizedExperiment(assays=list(counts=exp_data))
gnet_result <- gnet(se,reg_names,init_method,init_group_num)
plot_gene_group(gnet_result,group_idx=0)
}
