% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_group_correlation}
\alias{plot_group_correlation}
\title{Plot the correlation of each group}
\usage{
plot_group_correlation(gnet_result)
}
\arguments{
\item{gnet_result}{Results returned by gnet().}
}
\value{
A list of indices of the data point with correlation higher than the knee point.
}
\description{
Plot the correlation of each group and auto detected knee point. It can be used to 
determined which clustered are kept for further analysis.
}
\examples{
set.seed(1)
gnet_result <- list('group_score'=c(runif(10,1,3),c(runif(10,5,3))))
group_keep <- plot_group_correlation(gnet_result)
}
