% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnet_r.R
\name{plot_group_correlation}
\alias{plot_group_correlation}
\title{Plot the correlation of each group and auto detected knee point. It can be used to determined which clustered are kept for further analysis.}
\usage{
plot_group_correlation(gnet_result)
}
\arguments{
\item{avg_cor_list}{The average within group Pearson correlation coefficient of each group from gnet().}
}
\value{
A list of indices of the data point with correlation higher than the knee point.
}
\description{
Plot the correlation of each group and auto detected knee point. It can be used to determined which clustered are kept for further analysis.
}
