% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnet_r.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot the regression tree.}
\usage{
plot_tree(gnet_result, group_idx)
}
\arguments{
\item{gnet_result}{Results returned by gnet().}

\item{group_idx}{Index of the module..}
}
\description{
Plot the regression tree given the index of a module.
}
